build: preparation to switch to CSS only output
